Estimate significance of periodicity by permutation testing

DEPRECATED: We recommend instead using limorhyde, which has support for cosinor and periodic splines, in combination with methods such as limma.

zeitzeigerSig(x, time, nKnots = 3, nIter = 200, dopar = TRUE)

Arguments

x

Matrix of measurements, with observations in rows and features in columns. Missing values are allowed.

time

Vector of values of the periodic variable for the observations, where 0 corresponds to the lowest possible value and 1 corresponds to the highest possible value.

nKnots

Number of internal knots to use for the periodic smoothing spline.

nIter

Number of permutations.

dopar

Logical indicating whether to process features in parallel. Use doParallel::registerDoParallel() to register the parallel backend.

Value

Vector of p-values.

Details

Estimate the statistical significance of the periodic smoothing spline fit. At each permutation, the time vector is scrambled and then zeitzeigerFit() is used to fit a periodic smoothing spline for each feature as a function of time. The p-value for each feature is calculated based on the of permutations that had a signal-to-noise ratio at least as large as the observed signal-to-noise ratio, adjusted by the method of Phipson and Smyth (2010). Make sure to first register the parallel backend using doParallel::registerDoParallel(). For genome-scale data, this will be slow.

See also