Calculate sparse principal components of time-dependent variation on cross-validation

Calculate SPCs for each fold of cross-validation.

Usage

zeitzeigerSpcCv(
  fitResultList,
  nTime = 10,
  useSpc = TRUE,
  sumabsv = 1,
  orth = TRUE,
  dopar = TRUE
)

Arguments

fitResultList

Output of zeitzeigerFitCv().

nTime

Number of time-points by which to discretize the time-dependent behavior of each feature. Corresponds to the number of rows in the matrix for which the SPCs will be calculated.

useSpc

Logical indicating whether to use SPC (default) or svd.

sumabsv

L1-constraint on the SPCs, passed to SPC.

orth

Logical indicating whether to require left singular vectors be orthogonal to each other, passed to SPC.

dopar

Logical indicating whether to process the folds in parallel. Use doParallel::registerDoParallel() to register the parallel backend.

Value

A list consisting of the result from zeitzeigerSpc() for each fold.

See also